+++ src/components/Setup/context/SetupFlowContext.tsx
@@ -1,24 +1,5 @@
 import { createContext, useContext, useState, ReactNode } from 'react'
-
-export type ScreenId = 'welcome' | 'passkey' | 'add-wallets' | "contact-info" | 'success'
-export type ScreenProps = {
-    welcome: undefined
-    passkey: {
-        handle: string
-    }
-    'add-wallets': undefined
-    success: undefined
-    'contact-info': undefined
-}
-
-export type Step = {
-    screenId: ScreenId
-    title: string
-    description?: string
-    containerClassname: HTMLDivElement['className']
-    component: () => JSX.Element
-    centerComponent: () => JSX.Element
-}
+import { ScreenId, ScreenProps, Step } from '../Setup.types'
 
 interface SetupFlowContextType {
     currentStep: number
